# docs/api/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Bets","description":"Submit bets and evaluate jackpot outcomes"}],"paths":{"/api/v1/bets":{"post":{"tags":["Bets"],"summary":"Submit a bet","description":"Persists the bet and publishes it to Kafka for downstream processing.","operationId":"publishBet","requestBody":{"description":"Bet payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bet"},"examples":{"Sample bet":{"description":"Sample bet","value":{"betId":"b1","userId":"u1","jackpotId":"J1-FIXED","betAmount":75.0}}}}},"required":true},"responses":{"400":{"description":"Invalid request","content":{"application/problem+json":{}}},"200":{"description":"Bet accepted","content":{"text/plain":{"example":"Bet published to Kafka: b1"}}}}}},"/api/bets":{"post":{"tags":["Bets"],"summary":"Submit a bet","description":"Persists the bet and publishes it to Kafka for downstream processing.","operationId":"publishBet_1","requestBody":{"description":"Bet payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bet"},"examples":{"Sample bet":{"description":"Sample bet","value":{"betId":"b1","userId":"u1","jackpotId":"J1-FIXED","betAmount":75.0}}}}},"required":true},"responses":{"400":{"description":"Invalid request","content":{"application/problem+json":{}}},"200":{"description":"Bet accepted","content":{"text/plain":{"example":"Bet published to Kafka: b1"}}}}}},"/api/v1/jackpots/{betId}/evaluate":{"get":{"tags":["Bets"],"summary":"Evaluate jackpot outcome","description":"Returns the reward record for a bet; amount is 0.0 if not a winner.","operationId":"evaluate","parameters":[{"name":"betId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Invalid bet or jackpot","content":{"application/problem+json":{}}},"200":{"description":"Evaluation result","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reward"},"example":{"betId":"b1","userId":"u1","jackpotId":"J1-FIXED","jackpotRewardAmount":0.0,"createdAt":"2025-10-17T15:55:31.257Z"}}}}}}},"/api/jackpots/{betId}/evaluate":{"get":{"tags":["Bets"],"summary":"Evaluate jackpot outcome","description":"Returns the reward record for a bet; amount is 0.0 if not a winner.","operationId":"evaluate_1","parameters":[{"name":"betId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Invalid bet or jackpot","content":{"application/problem+json":{}}},"200":{"description":"Evaluation result","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reward"},"example":{"betId":"b1","userId":"u1","jackpotId":"J1-FIXED","jackpotRewardAmount":0.0,"createdAt":"2025-10-17T15:55:31.257Z"}}}}}}}},"components":{"schemas":{"Bet":{"required":["betId","jackpotId","userId"],"type":"object","properties":{"betId":{"type":"string","description":"Unique bet identifier","example":"b1"},"userId":{"type":"string","description":"User placing the bet","example":"u1"},"jackpotId":{"type":"string","description":"Target jackpot identifier","example":"J1-FIXED"},"betAmount":{"type":"number","description":"Bet amount (> 0)","format":"double","example":75.0}},"description":"Bet submission payload"},"Reward":{"type":"object","properties":{"betId":{"type":"string","description":"Evaluated bet id","example":"b1"},"userId":{"type":"string","description":"User id","example":"u1"},"jackpotId":{"type":"string","description":"Jackpot id","example":"J1-FIXED"},"jackpotRewardAmount":{"type":"number","description":"Payout amount (0.0 if not a winner)","format":"double","example":0.0},"createdAt":{"type":"string","description":"Creation timestamp (UTC ISO-8601)","format":"date-time","example":"2025-10-17T15:55:31.257Z"}},"description":"Jackpot evaluation result"}}}}